Movement Progress
So far, the basic function of player movement and jumping works, along with the camera following the player object on a 2D space.

So far so good. Now to find a way to implement a run button followed by gliding and zipline. And then after that to try applying everything to Catherine herself.

After that, adding in Bean and coding for that.