Finished with textures
Texturing, finished.
Now to add a night skybox, give some cel shading, add a little filter and a free cam. And then good to go.

And after that, I can add a little quick title screen as an extra bonus and stuff. Should be done by 12 if things go smoothly.